pyproject.toml
[tool.poetry]
name = "road-maintenance"
version = "0.1.0"
description = "Road Maintenance Management System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "road_maintenance"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.0.0"
django-environ = "1.0.0"
whitenoise = "6.6.0"
Pillow = "10.0.1"
psycopg2-binary = "2.9.9"
dj-database-url = "2.1.0"
djangorestframework = "3.15.1"
drf-spectacular = "0.28.0"
drf-spectacular-sidecar = "2024.4.1"
djangorestframework-simplejwt = "5.3.1"
django-allauth = "0.61.1"
python3-openid = "3.2.0"
PyJWT = "2.10.1"
channels = "4.1.0"
channels-redis = "4.2.0"
redis = "5.0.1"
django-crispy-forms = "2.1"
django-widget-tweaks = "1.5.0"
cryptography = "45.0.6"
python-dotenv = "1.0.0"
django-cors-headers = "4.3.1"
sendgrid = "6.11.0"
django-sendgrid-v5 = "1.2.0"
python-magic = "0.4.27"
django-cleanup = "8.0.0"
gunicorn = "21.2.0"
uvicorn = "0.29.0"
django-filter = "23.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
debug-toolbar = "6.0.0"
pytest = "^8.2.1"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
black = "^24.4.2"
flake8 = "^7.1.1"
isort = "^5.13.2"
